Minor changes to Lobby and new Log System
I made some minor changes to the Lobby class as usual, and I made a new
Log System for saving the logs. The log-system is going to be used
quite frequently so that we know what our customers in the future get
for errors. It's a very important feature to us, as it shows exactly
where the error occured. The line, the method-name and the class.